+++ packages/dodo-ui-bits/src/lib/components/Form/ToggleGroup/ToggleGroupSingle.svelte
@@ -1,14 +1,31 @@
 <script lang="ts">
   import { ToggleGroup as ToggleGroupBitUi, type SingleToggleGroupRootProps } from 'bits-ui';
 
+  type ToggleGroupSingleProps = {
+    allowDeselect?: boolean;
+  } & SingleToggleGroupRootProps;
+
   let {
     children,
     value = $bindable(),
     ref = $bindable(null),
+    allowDeselect = false,
     ...restProps
-  }: SingleToggleGroupRootProps = $props();
+  }: ToggleGroupSingleProps = $props();
+
+  function getValue() {
+    return value || '';
+  }
+
+  function setValue(newValue: string) {
+    if (!allowDeselect && !newValue) {
+      return;
+    }
+
+    value = newValue;
+  }
 </script>
 
-<ToggleGroupBitUi.Root {...restProps} bind:ref bind:value>
+<ToggleGroupBitUi.Root {...restProps} bind:ref bind:value={getValue, setValue}>
   {@render children?.()}
 </ToggleGroupBitUi.Root>